requirements.txt
# ===== Core =====
Django==5.2
Pillow>=10.4.0,<12.0.0

# ===== Security / Anti-spam =====
django-ratelimit==4.1.0
django-axes==6.5.1
django-honeypot==1.3.0

# ===== Production =====
whitenoise==6.7.0
python-dotenv==1.0.1
gunicorn==22.0.0
